pyproject.toml
[tool.poetry]
name = "demo-secret-manager"
version = "0.1.0"
description = "Demo GCP Secret manager"
authors = ["jggomezt <juan@wordbox.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-secret-manager = "^2.4.0"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
